backend/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
elasticsearch[async]
pydantic[email]
pydantic-settings
pytest
pytest-asyncio
pytest-cov
httpx
aiomysql
cryptography
